Move utils from ChatViewProvider
