+++ client/cody/src/__mocks__/vscode.js
@@ -1,4 +1,21 @@
 module.exports = {
-  workspace: { getConfiguration: () => undefined },
-  ConfigurationTarget: { Global: undefined },
+  window: {
+    showInformationMessage: () => undefined,
+    showErrorMessage: () => undefined,
+    showWarningMessage: () => undefined,
+    showQuickPick: () => undefined,
+    showInputBox: () => undefined,
+    createOutputChannel: () => undefined,
+  },
+  getConfiguration: () => undefined,
+  WorkspaceConfiguration: {
+    get: () => undefined,
+    update: () => undefined,
+  },
+  workspace: {
+    getConfiguration: () => undefined,
+  },
+  ConfigurationTarget: {
+    Global: undefined,
+  },
 }
